requirements.txt
Flask==2.3.3
PyMySQL==1.0.3
Flask-Cors==3.0.10
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
